94% test coverage

+++ se_leaverequestsystem/handlers/leave.py
@@ -1,19 +1,22 @@
 from datetime import datetime
 
-from flask import redirect, request, session, Response
+from flask import redirect, request, session
 
-from ..db.models import LeaveRequest
 from ..db import leave
+from ..db.models import LeaveRequest
 from ..extensions import db
 
 
-def deleteLeave(leave_id: int) -> str | dict[str, bool]:
-    leave_to_delete = LeaveRequest.query.get_or_404(leave_id)
+def deleteLeave(leave_id: int):
+
+    leave_to_delete = LeaveRequest.query.filter_by(id=leave_id).first()
+    if leave_to_delete is None:
+        return f"Leave with id {leave_id} does not exist", 404
 
     # Check if the logged-in user is the owner of the request
     if leave_to_delete.user_id == session.get("user_id"):
         if leave.startDatePassed(leave_to_delete.date_start):
-            return "Cannot delete because the start date has already passed"
+            return "Cannot delete because the start date has already passed", 400
 
         try:
             db.session.delete(leave_to_delete)
@@ -21,12 +24,12 @@ def deleteLeave(leave_id: int) -> str | dict[str, bool]:
             return {"delete_success": True}
         except Exception as e:
             print(f"Error: {e}")
-            return "There was an issue deleting your task"
+            return "There was an issue deleting your task", 500
     else:
-        return "You do not have permission to delete this request"
+        return "You do not have permission to delete this request", 403
 
 
-def postLeave() -> str | Response:
+def postLeave():
     user_id = session.get("user_id")
     leave_reason = request.form["reason"]
     leave_date_start_str = request.form["date_start"]
@@ -36,18 +39,18 @@ def postLeave() -> str | Response:
         leave_date_start = datetime.strptime(leave_date_start_str, "%Y-%m-%d")
         leave_date_end = datetime.strptime(leave_date_end_str, "%Y-%m-%d")
     except ValueError:
-        return "Please enter valid dates"
+        return "Please enter valid dates", 400
 
     if not leave.validatesSameDayConflict(leave_date_start):
-        return "Can't have 2 leaves starting on the same day"
+        return "Can't have 2 leaves starting on the same day", 400
 
     quota = 10
     if not leave.validatesLeaveQuota(leave_date_start, quota):
-        return f"You cannot have more than {quota} leaves in a year!"
+        return f"You cannot have more than {quota} leaves in a year!", 400
 
     max_days = 60
     if not leave.validatesMaxLeaveDate(leave_date_start, max_days):
-        return f"You cannot request leave over {max_days} days from now!"
+        return f"You cannot request leave over {max_days} days from now!", 400
 
     new_leave = LeaveRequest(
         reason=leave_reason,
@@ -62,4 +65,4 @@ def postLeave() -> str | Response:
         return redirect("/")
     except Exception as e:
         print(f"Error: {e}")
-        return "There was an issue adding your task"
+        return "There was an issue adding your task", 500
